eyeatt: RIP: it's all over.
Goodbye "eyeATT", it was nice to know you for a brief passing of time but
now is the time that you must go. 
Your use has expired as "eyeul" has grown all new powers 
(see cyanogenmod gerrit ID #101374) and will replace you.

The reason for your existance (firmware issues with the radio)
is over and you must be terminated.

You are now instructed to become a brand new powerful "eye".

Sorry that it wasn't working out and caused a lot of pain
in having to apply patches twice.

Good luck and enjoy your new life as part of the eyeul family.

Change-Id: Iff1e42434767d8ca700dc85f87d1649c7149a68f